refactor(payments-notifications-video): standardize domain module layout
