+++ src/screens/notifications/home/index.ts
@@ -1,0 +1,1 @@
+export { NotificationsScreen } from './view/NotificationsScreen';
