+++ src/screens/notifications/index.ts
@@ -1,0 +1,2 @@
+export { NotificationsScreen } from './home';
+export * from './routes';
