+++ src/screens/notifications/routes.ts
@@ -1,0 +1,3 @@
+export const NOTIFICATIONS_ROUTES = {
+  HOME: 'Notifications',
+} as const;
