+++ src/screens/payments/index.ts
@@ -1,0 +1,3 @@
+export { PaymentsScreen } from './payments';
+export { WalletLedgerScreen } from './wallet';
+export * from './routes';
