+++ src/screens/payments/payments/index.ts
@@ -1,0 +1,1 @@
+export { PaymentsScreen } from './view/PaymentsScreen';
