+++ src/screens/payments/wallet/index.ts
@@ -1,0 +1,1 @@
+export { WalletLedgerScreen } from './view/WalletLedgerScreen';
