+++ src/screens/video/reels/index.ts
@@ -1,0 +1,1 @@
+export { SoulReelsScreen } from './view/SoulReelsScreen';
