+++ src/screens/video/routes.ts
@@ -1,0 +1,3 @@
+export const VIDEO_ROUTES = {
+  SOUL_REELS: 'SoulReels',
+} as const;
